Attempt at fixing sandboxing/security flaw
Also cleaned up the directories, hopefully uploaded files aren't kept
around when no longer needed. This should fix any cross-compatibility
issues with different server configurations.

Hopefully it also stops python from accessing the root directory of the
server drive, instead accessing flasks root - which may be different on
different configurations.

Still need to investigate how secure this is with more testing, but it's
a hell of a lot more secure than it was.